fix: inverted logic to get push/pull strings
the functions to get the git command to count the number of commits to
push/pull were inverted when a branch was passed in. The push command
counted the number of commits to pull, and vice versa. Fixed that, and
made the branch ui push/pull string highlighting. match the status ui.